[#20250104] Move app init logic to dedicated module
* Extracts initialization logic from main.ts
* Creates dedicated appInitializer Vue plugin
* Moves startup banner to logging service for better separation
* Simplifies main.ts entry point to focus on core app creation
* Adds type-safe configuration options for initialization
