+++ src/services/logging.ts
@@ -1,10 +1,17 @@
 // services/logging.ts
 
+const STARTUP_BANNER = `
+┏┓┳┓┏┓┏┳┓┳┳┳┓┏┓
+┃┃┃┃┣  ┃ ┃┃┃┃┣
+┗┛┛┗┗┛ ┻ ┻┛ ┗┗┛
+`;
+
 export interface LoggingService {
   error(error: Error): void;
   warn(message: string, context?: Record<string, unknown>): void;
   info(message: string, context?: Record<string, unknown>): void;
   debug(message: string, context?: Record<string, unknown>): void;
+  banner(): void;
 }
 
 // services/logging/console.ts
@@ -26,6 +33,10 @@ export class ConsoleLoggingService implements LoggingService {
 
   debug(message: string, context?: Record<string, unknown>): void {
     console.debug(message, context);
+  }
+
+  banner(): void {
+    console.log(STARTUP_BANNER);
   }
 }
